% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hflm.R
\name{create.lag.matrix}
\alias{create.lag.matrix}
\title{Indicator matrix for Lag and Historical FLMs}
\usage{
create.lag.matrix(delta, taus, S, ntaus)
}
\arguments{
\item{delta}{post burn-in MCMC samples from `bayeshflm()`}

\item{taus}{the credibility level required}

\item{S}{s specification of which type of credibility interval should be generates. Only pointwise or joint CIs are supported.}

\item{ntaus}{s specification of which type of credibility interval should be generates. Only pointwise or joint CIs are supported.}
}
\value{
A matrix with columns giving lower and upper credibility limits for each parameter.
}
\description{
Produces an indicator matrix for the H/LFLMs
}
