% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hflm.R
\name{predict_hflm}
\alias{predict_hflm}
\title{Credibility intervals for estimates of the parameters of the Bayes HFLM}
\usage{
predict_hflm(hflm_mcmc, X1_new, X2_new, taus, DEL, interval = TRUE)
}
\arguments{
\item{mcmc_samples}{post burn-in MCMC samples from \code{\link{bayeshflm}}}

\item{level}{the credibility level required}

\item{type}{s specification of which type of credibility interval should be generates. Only pointwise or joint CIs are supported.}
}
\value{
A matrix with columns giving lower and upper credibility limits for each parameter.
}
\description{
Produces joint or pointwise credibility intervals of the functional parameters of the model \code{\link{brocolors}}
}
