% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hflm.R
\name{predict_hflm_joint}
\alias{predict_hflm_joint}
\title{Generate predictions from the predictive posterior distribution for the Bayes HFLM}
\usage{
predict_hflm_joint(
  hflm_mcmc,
  X1_new,
  X2_new,
  taus,
  DEL,
  burnin,
  interval = TRUE,
  level = 0.95
)
}
\arguments{
\item{level}{the credibility level required}

\item{mcmc_samples}{post burn-in MCMC samples from \code{\link{bayeshflm}}}

\item{type}{s specification of which type of credibility interval should be generates. Only pointwise or joint CIs are supported.}
}
\value{
A matrix with columns giving lower and upper credibility limits for each parameter.
}
\description{
Produces the predictions and their prediction intervals from new data set.
}
