% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_lflm.R
\name{simulate.hflm}
\alias{simulate.hflm}
\title{Credibility intervals for estimates of the parameters of the Bayes HFLM}
\usage{
\method{simulate}{hflm}(
  nobs = 100,
  n.tau = 25,
  beta1 = "bimodal-lagged",
  varx1 = 0.1,
  beta2 = "null",
  varx2 = 0.1,
  eSNR = 5,
  delta = NULL,
  plot = TRUE,
  seed = 1234
)
}
\arguments{
\item{mcmc_samples}{post burn-in MCMC samples from \code{\link{bayeshflm}}}

\item{level}{the credibility level required}

\item{type}{s specification of which type of credibility interval should be generates. Only pointwise or joint CIs are supported.}
}
\value{
A matrix with columns giving lower and upper credibility limits for each parameter.
}
\description{
Produces joint or pointwise credibility intervals of the functional parameters of the model \code{\link{bayeshflm}}
}
