% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hflm.R
\name{bayeshflm}
\alias{bayeshflm}
\title{Credibility intervals for estimates of the parameters of the Bayes HFLM}
\usage{
bayeshflm(
  Ymat,
  Xmat1,
  Xmat2,
  taus,
  Ss,
  interceptInfo = list(Fk1 = "bs", K1 = 10),
  tensorInfo = list(Fk = "bs", Psi = "bs", U = 10, K = 10),
  fpcaInfo = list(knots = 5, npc = tensorInfo$U),
  lag = NULL,
  niter = 1000,
  nburn = 0.5,
  mcmc_hyper = list(i1 = 0.001, i2 = 0.001, a = 0.001, b = 0.001, A = 1, B = 1, C = 1),
  diagonalise = TRUE,
  smooth = TRUE,
  prior = "HS"
)
}
\arguments{
\item{mcmc_samples}{post burn-in MCMC samples from `bayeshflm()`}

\item{level}{the credibility level required}

\item{type}{s specification of which type of credibility interval should be generates. Only pointwise or joint CIs are supported.}
}
\value{
A matrix with columns giving lower and upper credibility limits for each parameter.
}
\description{
Produces joint or pointwise credibility intervals of the functional parameters of the model \code{\link{brocolors}}
}
